Arpeggiators: Handle CC 123 by emptying the chord and latched notes cache.
This helps to get rid of hanging notes due to missing note offs in the
MIDI input. The cause of these is unknown, so we haven't been able to
resolve that bug yet, but at least you can now get rid of the hanging
notes by hitting Ardour's panic button.