Add raptor_arp.lua, a quick-and-dirty port of the random arpeggiator.
This is still a bit experimental, so expect some bugs (and please report
them). Some factory presets from the Pd version of this program are
included for illustration purposes.

Raptor is based on Clarence Barlow's theories on harmony and meter.
It's really a full-blown algorithmic composition tool in the form of an
arpeggiator, although it offers the usual run-of-the-mill deterministic
and random arpeggios as well.

The real magic starts when you turn on raptor mode and start playing
around with some of the controls. With the right choice of parameters,
Raptor can go from plain tonal to a more jazz-like feel to completely
atonal (or even anti-tonal) in a continuous fashion. Such "harmonicity
sweeps" are a hallmark feature of many of Barlow's compositions. With
Raptor you can "go Barlow" on a whim and quickly return to the safe
harbor of tonality, or play an entire piece in (an approximation of)
"Barlow style" with varying degrees of tonality.

Ardour-specific documentation still needs to be written, but see
https://github.com/agraef/raptor-lua for the Pd version of this program,
which offers extensive documentation of the algorithm and its features.