raptor_arp: Overhaul of the control values.
Change fractions (stuff like gate, probability, and harmonicity values;
basically, everything that's not a MIDI-related parameter) to normalized
values in the -1..1 or 0..1 range. This is for consistency with the
other arp plugins.

We inherited the previous -100..100 and 0..100 ranges from Pd Raptor
which uses percentage values for these parameters, but implemented them
as integers, loosing precision, so the new scheme also offers better
resolution.