raptor_arp: Work around lost note-offs at loop end with automation (again).
This issue keeps coming back and is rather unpredictable (I have
sessions working without a hitch on one machine and giving me hanging
notes on another). So we just keep regular_block_length enabled for now,
which seems to completely get rid of the problem. Thanks to Robin Gareus
for adding this option! (Only in git at the time of this writing.)